Created a context file for submit-link
